+++ common/src/main/java/net/kyronis/better_mcdonalds_mod/common/BetterMcDonaldsMod.java
@@ -11,7 +11,6 @@ public class BetterMcDonaldsMod {
         registrars.blocks(BMMBlocks::initialize);
         registrars.items(BMMItems::initialize);
         registrars.creativeModeTabs(BMMCreativeTabs::initialize);
-        registrars.compostables(BMMCompostables::initialize);
         BMMLootTables.initialize(Balm.lootModifiers());
         BMMWorldGen.initialize(Balm.biomeModifications());
     }