+++ common/src/main/java/net/kyronis/better_mcdonalds_mod/common/loot/modifier/BMMSeedLootModifier.java
@@ -26,7 +26,7 @@ public class BMMSeedLootModifier implements BalmLootModifier {
             return;
         }
 
-        BlockState state = context.getOptionalParameter(LootContextParams.BLOCK_STATE);
+        BlockState state = context.getOptional(LootContextParams.BLOCK_STATE);
 
         if (!GRASS_BLOCKS.contains(state.getBlock())) return;