+++ common/src/main/java/net/kyronis/better_mcdonalds_mod/common/worldgen/placed_feature/BMMPlacedFeatures.java
@@ -1,7 +1,7 @@
 package net.kyronis.better_mcdonalds_mod.common.worldgen.placed_feature;
 
 import net.kyronis.better_mcdonalds_mod.common.BetterMcDonaldsMod;
-import net.kyronis.better_mcdonalds_mod.common.worldgen.configured_feature.BMMConfiguredFeatures;
+import net.kyronis.better_mcdonalds_mod.common.worldgen.feature.BMMFeatures;
 import net.minecraft.core.HolderGetter;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.data.worldgen.BootstrapContext;
@@ -9,23 +9,23 @@ import net.minecraft.resources.Identifier;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.level.levelgen.Heightmap;
 import net.minecraft.world.level.levelgen.blockpredicates.BlockPredicate;
-import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
+import net.minecraft.world.level.levelgen.feature.Feature;
 import net.minecraft.world.level.levelgen.placement.*;
 import net.minecraft.world.level.material.Fluids;
 
 import java.util.List;
 
 public class BMMPlacedFeatures {
-    public static final ResourceKey<PlacedFeature> SALT_BLOCK = registerKey("salt_block");
+    public static final ResourceKey<PlacedFeature> DISK_SALT = registerKey("disk_salt");
 
     public static void bootstrap(BootstrapContext<PlacedFeature> context) {
-        HolderGetter<ConfiguredFeature<?, ?>> configuredFeatures = context.lookup(Registries.CONFIGURED_FEATURE);
+        HolderGetter<Feature> features = context.lookup(Registries.FEATURE);
 
-        context.register(SALT_BLOCK, saltBlockFeature(configuredFeatures));
+        context.register(DISK_SALT, saltBlockFeature(features));
     }
 
-    private static PlacedFeature saltBlockFeature(HolderGetter<ConfiguredFeature<?, ?>> configuredFeatures) {
-        return new PlacedFeature(configuredFeatures.getOrThrow(BMMConfiguredFeatures.SALT_BLOCK), List.of(
+    private static PlacedFeature saltBlockFeature(HolderGetter<Feature> features) {
+        return new PlacedFeature(features.getOrThrow(BMMFeatures.DISK_SALT), List.of(
                 CountPlacement.of(2),
                 InSquarePlacement.spread(),
                 HeightmapPlacement.onHeightmap(Heightmap.Types.OCEAN_FLOOR_WG),