+++ fabric/src/main/java/net/kyronis/better_mcdonalds_mod/datagen/provider/BMMRecipeProvider.java
@@ -1,13 +1,16 @@
 package net.kyronis.better_mcdonalds_mod.datagen.provider;
 
+import net.fabricmc.fabric.api.datagen.v1.FabricPackOutput;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricRecipeProvider;
 import net.kyronis.better_mcdonalds_mod.common.BetterMcDonaldsMod;
 import net.kyronis.better_mcdonalds_mod.common.registry.BMMBlocks;
 import net.kyronis.better_mcdonalds_mod.common.registry.BMMItems;
 import net.kyronis.better_mcdonalds_mod.common.tags.BMMTags;
+import net.minecraft.advancements.Advancement;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.core.registries.Registries;
-import net.minecraft.data.PackOutput;
 import net.minecraft.data.recipes.*;
+import net.minecraft.data.worldgen.BootstrapContext;
 import net.minecraft.resources.Identifier;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.item.Item;
@@ -19,14 +22,14 @@ import net.minecraft.world.level.ItemLike;
 
 import java.util.concurrent.CompletableFuture;
 
-public class BMMRecipeProvider extends RecipeProvider.Runner {
-    public BMMRecipeProvider(PackOutput packOutput, CompletableFuture<HolderLookup.Provider> registries) {
-        super(packOutput, registries);
+public class BMMRecipeProvider extends FabricRecipeProvider {
+    public BMMRecipeProvider(FabricPackOutput output, CompletableFuture<HolderLookup.Provider> registriesFuture) {
+        super(output, registriesFuture);
     }
 
     @Override
-    protected RecipeProvider createRecipeProvider(HolderLookup.Provider provider, RecipeOutput recipeOutput) {
-        return new RecipeProvider(provider, recipeOutput) {
+    protected RecipeProvider createRecipeProvider(HolderLookup.Provider registries, BootstrapContext<Recipe<?>> recipes, BootstrapContext<Advancement> advancements) {
+        return new RecipeProvider(recipes, advancements) {
             @Override
             public void buildRecipes() {
                 HolderLookup.RegistryLookup<Item> itemLookup = registries.lookupOrThrow(Registries.ITEM);
@@ -315,11 +318,5 @@ public class BMMRecipeProvider extends RecipeProvider.Runner {
                 return ResourceKey.create(Registries.RECIPE, Identifier.fromNamespaceAndPath(BetterMcDonaldsMod.MOD_ID, getItemName(item)));
             }
         };
-
-    }
-
-    @Override
-    public String getName() {
-        return BetterMcDonaldsMod.MOD_ID + " recipes";
     }
 }