+++ fabric/src/main/java/net/kyronis/better_mcdonalds_mod/datagen/provider/BMMRegistryDataProvider.java
@@ -1,0 +1,26 @@
+package net.kyronis.better_mcdonalds_mod.datagen.provider;
+
+import net.fabricmc.fabric.api.datagen.v1.FabricPackOutput;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricDynamicRegistryProvider;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.core.registries.Registries;
+
+import java.util.concurrent.CompletableFuture;
+
+public class BMMRegistryDataProvider extends FabricDynamicRegistryProvider {
+    public BMMRegistryDataProvider(FabricPackOutput output, CompletableFuture<HolderLookup.Provider> registriesFuture) {
+        super(output, registriesFuture);
+    }
+
+    @Override
+    protected void configure(HolderLookup.Provider registries, Entries entries) {
+        entries.addAll(registries.lookupOrThrow(Registries.FEATURE));
+        entries.addAll(registries.lookupOrThrow(Registries.PLACED_FEATURE));
+        entries.addAll(registries.lookupOrThrow(Registries.VILLAGER_TRADE));
+    }
+
+    @Override
+    public String getName() {
+        return "";
+    }
+}