+++ changelog.md
@@ -1,3 +1,10 @@
 <h2>Changes</h2>
 
-- Implement new NeoForge mods list screen features
+- Updated to Minecraft 26.3
+- Bump version
+
+<h2>Technical Changes</h2>
+
+- Added `disk_salt` Feature
+- Renamed `salt_block` Placed Feature to `disk_salt`
+- Removed `salt_block` Configured Feature